+++ src/test/java/com/project/singk/mock/FakeActivityHistoryRepository.java
@@ -2,6 +2,7 @@ package com.project.singk.mock;
 
 import com.project.singk.domain.activity.domain.ActivityHistory;
 import com.project.singk.domain.activity.service.port.ActivityHistoryRepository;
+import org.springframework.data.domain.Page;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -16,6 +17,16 @@ public class FakeActivityHistoryRepository implements ActivityHistoryRepository 
         data.removeIf(item -> Objects.equals(item.getId(), activityHistory.getId()));
         data.add(activityHistory);
         return activityHistory;
+    }
+
+    @Override
+    public List<ActivityHistory> getActivityGraph(Long memberId) {
+        return null;
+    }
+
+    @Override
+    public Page<ActivityHistory> getActivityHistories(Long memberId, int offset, int limit) {
+        return null;
     }
 
 }
